fix:修改打印

+++ core/src/config/plugins/removeStyle.ts
@@ -1,12 +1,15 @@
-import { OutputOptions } from 'rollup';
 import path from 'node:path';
 import FS from 'fs-extra';
-export function removeStyle() {
+import type { Plugin } from 'vite';
+
+export function removeStyle(): Plugin {
   return {
     name: 'saqu:remove-style',
-    writeBundle(options: OutputOptions, bundle: any) {
-      if (bundle && bundle['style.css']) {
-        if (/^export default undefined;/.test(bundle['style.css'].source) || !bundle['style.css'].source) {
+    writeBundle(options, bundle) {
+      if (bundle && bundle['style.css'] && (bundle['style.css'] as any).source) {
+        const content = (bundle['style.css'] as any).source;
+        if (/^export default undefined;/.test(content) || !content) {
+          delete bundle['style.css'];
           const stylePath = path.resolve(process.cwd(), options.dir, 'style.css');
           if (FS.existsSync(stylePath)) {
             FS.removeSync(stylePath);
